@@ -1,12 +1,11 @@
-// swift-tools-version: 5.10
+// swift-tools-version: 6.0
 // The swift-tools-version declares the minimum version of Swift required to build this package.
 
 import PackageDescription
 
 let settings: [SwiftSetting] = [
     .unsafeFlags(["-warnings-as-errors"]),
-    .enableExperimentalFeature("ExistentialAny"),
-    .enableExperimentalFeature("StrictConcurrency=targeted")
+    .enableExperimentalFeature("ExistentialAny")
 ]
 
 let package = Package(
@@ -57,5 +56,6 @@ let package = Package(
             ],
             swiftSettings: settings
         )
-    ]
+    ],
+    swiftLanguageModes: [.v6]
 )
